Merge with master

+++ ZeroMev/MevEFC/Token.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 
 namespace ZeroMev.MevEFC
 {
     public partial class Token
     {
         public string TokenAddress { get; set; } = null!;
-        public decimal Decimals { get; set; }
+        public BigInteger Decimals { get; set; }
     }
 }
